# docs/reference/api/openapi.yaml
openapi: 3.1.0
info:
  title: Capacitarr API
  version: 3.0.0
  description: |
    Capacitarr is a capacity management tool for media servers. It monitors disk
    usage across Sonarr, Radarr, Lidarr, Readarr, Plex, Jellyfin, Emby, Tautulli,
    and Seerr integrations, scores media items for potential cleanup based on
    configurable weights and rules, and can automatically (or via approval) free
    disk space when thresholds are breached.

    ## Authentication

    Most endpoints require authentication via one of:
    - **JWT cookie** (`jwt`) — set automatically on login
    - **Bearer token** — `Authorization: Bearer <token>` header
    - **API key** — `X-Api-Key: <key>` header

    The `/health`, `/version`, `/auth/login`, and `/auth/logout` endpoints are public.
  license:
    name: PolyForm Noncommercial 1.0.0
    url: https://polyformproject.org/licenses/noncommercial/1.0.0/
servers:
  - url: http://localhost:2187/api/v1
    description: Local development server
security:
  - cookieAuth: []
  - bearerAuth: []
  - apiKey: []
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [Auth]
      summary: Log in
      description: |
        Authenticate with username and password. Returns an HttpOnly `jwt` cookie
        and a non-HttpOnly `authenticated` indicator cookie. On first call with no
        existing users, the provided credentials bootstrap the initial admin user.
        Rate-limited to 10 attempts per IP per 15-minute window.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              description: HttpOnly JWT cookie and `authenticated=true` indicator cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /auth/password:
    put:
      operationId: changePassword
      tags: [Auth]
      summary: Change password
      description: Change the current user's password. Requires the current password for verification. New password must be at least 8 characters.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Current password is incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/username:
    put:
      operationId: changeUsername
      tags: [Auth]
      summary: Change username
      description: Change the current user's username. Requires the current password for verification. New username must be at least 3 characters and not already taken.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeUsernameRequest"
      responses:
        "200":
          description: Username changed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Current password is incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/apikey:
    get:
      operationId: getApiKeyStatus
      tags: [Auth]
      summary: Check API key status
      description: Returns whether the current user has an API key configured. Never returns the actual key.
      responses:
        "200":
          description: API key status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: generateApiKey
      tags: [Auth]
      summary: Generate a new API key
      description: |
        Generates a new 256-bit random API key. The plaintext key is returned
        **once** in the response and is never stored — only a SHA-256 hash is
        persisted. Any previously existing key is replaced.
      responses:
        "200":
          description: New API key generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ===========================================================================
  # Metrics
  # ===========================================================================
...
components:
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key generated via `POST /auth/apikey`
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login response
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
      description: HttpOnly JWT cookie set on login

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
...
  responses:
    BadRequest:
      description: Bad request — invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
      required: [error]
      example:
        error: "Invalid request body"
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
      example:
        message: "Password changed successfully"

    # ---- Health / Version ----
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
          example: mypassword
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        message:
          type: string
          example: success
      required: [message]
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
          minLength: 8
      required: [currentPassword, newPassword]
    ChangeUsernameRequest:
      type: object
      properties:
        newUsername:
          type: string
          minLength: 3
        currentPassword:
          type: string
          format: password
      required: [newUsername, currentPassword]
    ApiKeyStatusResponse:
      type: object
      properties:
        has_key:
          type: boolean
          description: Whether the user has an API key configured
      required: [has_key]
      example:
        has_key: true
    ApiKeyResponse:
      type: object
      properties:
        api_key:
          type: string
          description: |
            The plaintext API key. This is shown **once** and never stored.
            Save it securely.
          example: "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2"
      required: [api_key]

    # ---- Disk Group ----
...